refactored query variables from view
