+++ packages/twenty-front/src/modules/object-record/record-show/hooks/useRecordShowPagePagination.ts
@@ -1,5 +1,4 @@
 /* eslint-disable @nx/workspace-no-navigate-prefer-link */
-import { useMemo } from 'react';
 import {
   useLocation,
   useNavigate,
@@ -9,42 +8,15 @@ import {
 import { useSetRecoilState } from 'recoil';
 
 import { useObjectMetadataItem } from '@/object-metadata/hooks/useObjectMetadataItem';
-import { FieldMetadataItem } from '@/object-metadata/types/FieldMetadataItem';
-import { ObjectMetadataItem } from '@/object-metadata/types/ObjectMetadataItem';
-import { formatFieldMetadataItemsAsFilterDefinitions } from '@/object-metadata/utils/formatFieldMetadataItemsAsFilterDefinitions';
-import { formatFieldMetadataItemsAsSortDefinitions } from '@/object-metadata/utils/formatFieldMetadataItemsAsSortDefinitions';
 import { generateDepthOneRecordGqlFields } from '@/object-record/graphql/utils/generateDepthOneRecordGqlFields';
 import { useFindManyRecords } from '@/object-record/hooks/useFindManyRecords';
-import { turnSortsIntoOrderBy } from '@/object-record/object-sort-dropdown/utils/turnSortsIntoOrderBy';
 import { lastShowPageRecordIdState } from '@/object-record/record-field/states/lastShowPageRecordId';
-import { turnObjectDropdownFilterIntoQueryFilter } from '@/object-record/record-filter/utils/turnObjectDropdownFilterIntoQueryFilter';
 import { useRecordIdsFromFindManyCacheRootQuery } from '@/object-record/record-show/hooks/useRecordIdsFromFindManyCacheRootQuery';
 import { buildShowPageURL } from '@/object-record/record-show/utils/buildShowPageURL';
 import { buildIndexTablePageURL } from '@/object-record/record-table/utils/buildIndexTableURL';
-import { usePrefetchedData } from '@/prefetch/hooks/usePrefetchedData';
-import { PrefetchKey } from '@/prefetch/types/PrefetchKey';
-import { View } from '@/views/types/View';
-import { mapViewFiltersToFilters } from '@/views/utils/mapViewFiltersToFilters';
-import { mapViewSortsToSorts } from '@/views/utils/mapViewSortsToSorts';
-import { filter } from '@chakra-ui/system';
+import { useQueryVariablesFromActiveFieldsOfViewOrDefaultView } from '@/views/hooks/useQueryVariablesFromActiveFieldsOfViewOrDefaultView';
 import { isNonEmptyString } from '@sniptt/guards';
 import { capitalize } from '~/utils/string/capitalize';
-
-export const findView = ({
-  viewId,
-  objectMetadataItemId,
-  views,
-}: {
-  viewId: string | null;
-  objectMetadataItemId: string;
-  views: View[];
-}) => {
-  return views.find(
-    (view: View) =>
-      (view.key === 'INDEX' || view?.id === viewId) &&
-      view?.objectMetadataId === objectMetadataItemId,
-  );
-};
 
 export const useRecordShowPagePagination = (
   propsObjectNameSingular: string,
@@ -69,77 +41,6 @@ export const useRecordShowPagePagination = (
 
   const { objectMetadataItem } = useObjectMetadataItem({ objectNameSingular });
 
-  const useViewFromPrefetchedViews = ({
-    objectMetadataItemId,
-    viewId,
-  }: {
-    objectMetadataItemId: string;
-    viewId: string | null;
-  }) => {
-    const { records: views } = usePrefetchedData<View>(PrefetchKey.AllViews);
-
-    const view = useMemo(() => {
-      return findView({
-        objectMetadataItemId,
-        viewId,
-        views,
-      });
-    }, [viewId, views, objectMetadataItemId]);
-
-    return view;
-  };
-
-  const useActiveFieldMetadataItems = ({
-    objectMetadataItem,
-  }: {
-    objectMetadataItem: ObjectMetadataItem;
-  }) => {
-    const activeFieldMetadataItems = useMemo(
-      () =>
-        objectMetadataItem
-          ? objectMetadataItem.fields.filter(
-              ({ isActive, isSystem }) => isActive && !isSystem,
-            )
-          : [],
-      [objectMetadataItem],
-    );
-
-    return activeFieldMetadataItems;
-  };
-
-  const getQueryVariablesFromView = ({
-    view,
-    fieldMetadataItems,
-  }: {
-    view: View;
-    fieldMetadataItems: FieldMetadataItem[];
-  }) => {
-    const { viewFilters, viewSorts } = view;
-
-    const filterDefinitions = formatFieldMetadataItemsAsFilterDefinitions({
-      fields: fieldMetadataItems,
-    });
-
-    const sortDefinitions = formatFieldMetadataItemsAsSortDefinitions({
-      fields: fieldMetadataItems,
-    });
-
-    const filter = turnObjectDropdownFilterIntoQueryFilter(
-      mapViewFiltersToFilters(viewFilters, filterDefinitions),
-      objectMetadataItem?.fields ?? [],
-    );
-
-    const orderBy = turnSortsIntoOrderBy(
-      objectMetadataItem,
-      mapViewSortsToSorts(viewSorts, sortDefinitions),
-    );
-
-    return {
-      filter,
-      orderBy,
-    };
-  };
-
   const recordGqlFields = generateDepthOneRecordGqlFields({
     objectMetadataItem,
   });
@@ -147,6 +48,12 @@ export const useRecordShowPagePagination = (
   const { state } = useLocation();
 
   const cursorFromIndexPage = state?.cursor;
+
+  const { filter, orderBy } =
+    useQueryVariablesFromActiveFieldsOfViewOrDefaultView({
+      objectMetadataItem,
+      viewId: viewIdQueryParam,
+    });
 
   const { loading: loadingCurrentRecord, pageInfo: currentRecordsPageInfo } =
     useFindManyRecords({
@@ -255,22 +162,6 @@ export const useRecordShowPagePagination = (
       orderBy,
     },
   });
-
-  const getShowPageTitle = () => {
-    const rankInView = recordIdsInCache.findIndex(
-      (id) => id === objectRecordId,
-    );
-
-    const rankFoundInFiew = rankInView > -1;
-
-    const objectLabel = capitalize(objectMetadataItem.namePlural);
-
-    const viewNameWithCount = rankFoundInFiew
-      ? `${rankInView + 1} of ${totalCount} in ${objectLabel}`
-      : `${objectLabel} (${totalCount})`;
-
-    return viewNameWithCount;
-  };
 
   const rankInView = recordIdsInCache.findIndex((id) => id === objectRecordId);
