+++ packages/twenty-front/src/modules/views/hooks/useQueryVariablesFromActiveFieldsOfViewOrDefaultView.ts
@@ -1,0 +1,32 @@
+import { useActiveFieldMetadataItems } from '@/object-metadata/hooks/useActiveFieldMetadataItems';
+import { ObjectMetadataItem } from '@/object-metadata/types/ObjectMetadataItem';
+import { useViewOrDefaultViewFromPrefetchedViews } from '@/views/hooks/useViewOrDefaultViewFromPrefetchedViews';
+import { getQueryVariablesFromView } from '@/views/utils/getQueryVariablesFromView';
+
+export const useQueryVariablesFromActiveFieldsOfViewOrDefaultView = ({
+  objectMetadataItem,
+  viewId,
+}: {
+  objectMetadataItem: ObjectMetadataItem;
+  viewId: string | null | undefined;
+}) => {
+  const { view } = useViewOrDefaultViewFromPrefetchedViews({
+    objectMetadataItemId: objectMetadataItem.id,
+    viewId,
+  });
+
+  const { activeFieldMetadataItems } = useActiveFieldMetadataItems({
+    objectMetadataItem,
+  });
+
+  const { filter, orderBy } = getQueryVariablesFromView({
+    fieldMetadataItems: activeFieldMetadataItems,
+    objectMetadataItem,
+    view,
+  });
+
+  return {
+    filter,
+    orderBy,
+  };
+};
